examples/research/comcast-anomaly/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comcast-anomaly-fl"
version = "0.1.0"
description = "Flower FL pipeline for Comcast anomaly synthetic v2 models"
license = "Apache-2.0"
dependencies = [
  "flwr[simulation]>=1.26.1",
  "numpy>=1.24",
  "torch>=2.0",
  "scikit-learn>=1.3",
  "pyyaml>=6.0",
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "jinja2>=3.1",
]

[tool.hatch.build.targets.wheel]
packages = ["comcast_fl", "comcast_ui"]

[tool.flwr.app]
publisher = "micahsheller"

[tool.flwr.app.components]
serverapp = "comcast_fl.flower_server_app:app"
clientapp = "comcast_fl.flower_client_app:app"

[tool.flwr.app.config]
experiment-config-json = "{}"
domain = "downstream_rxmer"
run-name = "default_run"
output-root = "artifacts/fl"
